+++ src/main/java/net/brdle/collectorsreap/common/fluid/DrinkFluidType.java
@@ -7,10 +7,10 @@ import net.minecraftforge.common.SoundActions;
 import net.minecraftforge.fluids.FluidType;
 import java.util.function.Consumer;
 
-public class TeaFluidType extends FluidType {
+public class DrinkFluidType extends FluidType {
 	private final int tint;
 
-	public TeaFluidType(int tint) {
+	public DrinkFluidType(int tint) {
 		super(Properties.create()
 			.sound(SoundActions.BUCKET_FILL, SoundEvents.BUCKET_FILL)
 			.sound(SoundActions.BUCKET_EMPTY, SoundEvents.BUCKET_EMPTY)