+++ src/main/java/net/brdle/collectorsreap/data/gen/CRLanguageProvider.java
@@ -42,6 +42,7 @@ public class CRLanguageProvider extends LanguageProvider {
 		add("trim_material." + CollectorsReap.MODID + ".pearl", "Pearl Material");
 		add("fluid_type." + CollectorsReap.MODID + ".lime_green_tea_type", "Lime Green Tea");
 		add("fluid_type." + CollectorsReap.MODID + ".pomegranate_black_tea_type", "Pomegranate Black Tea");
+		add("fluid_type." + CollectorsReap.MODID + ".deific_blood_type", "Deific Blood");
 		add("tooltip.requires_modid", "Requires modid:");
 		add("desc." + CollectorsReap.MODID + "." + Util.name(CRItems.PORTOBELLO), "Portobellos can be rarely found in meadows.");
 		add("desc." + CollectorsReap.MODID + "." + Util.name(CRItems.LIME), "Limes can be found as a wild bush in flower forests.");
@@ -158,6 +159,7 @@ public class CRLanguageProvider extends LanguageProvider {
 		addItem(CRItems.LIME_GREEN_TEA, "Lime Green Tea");
 		addItem(CRItems.POMEGRANATE_BLACK_TEA, "Pomegranate Black Tea");
 		addItem(CRItems.POMEGRANATE_SMOOTHIE, "Pomegranate Smoothie");
+		addItem(CRItems.DEIFIC_BLOOD, "Deific Blood");
 		addItem(CRItems.POMEGRANATE, "Pomegranate");
 		addItem(CRItems.POMEGRANATE_SEEDS, "Pomegranate Seeds");
 		addItem(CRItems.POMEGRANATE_SLICE, "Pomegranate Slice");